requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
gotify==0.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
requests==2.31.0
sniffio==1.3.1
typing_extensions==4.10.0
urllib3==2.2.1
websockets==12.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
gotify==0.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
typing_extensions==4.10.0
urllib3==2.2.1
websockets==12.0
